fix: removed code duplication from account processing
