+++ src/acct/tests.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
     use crate::acct::account::{self, Account};
-    use crate::acct::storage;
+    use crate::acct::storage::{self, AccountStorage};
     use crate::tx::{storage as TxStore, transaction::Transaction};
 
     #[test]
@@ -1128,6 +1128,71 @@ mod tests {
             "wrong locked status; expect {}, got {}",
             true,
             account.locked
+        );
+    }
+
+    #[test]
+    fn test_account_storage_insert() {
+        let account = Account {
+            client: 1,
+            available: 20.0,
+            held: 0.0,
+            total: 0.0,
+            locked: false,
+        };
+
+        let db = AccountStorage::new();
+        let acct_exist = db.exists(1);
+        assert!(!acct_exist, "account should be empty");
+
+        db.insert(account.clone());
+
+        let acct: Account = db.read(1, |acct| acct.unwrap().clone());
+        // println!("{:?}", acct);
+
+        assert!(
+            acct.client == account.client,
+            "created client id and fetched client id are not equal"
+        );
+    }
+
+    #[test]
+    fn test_account_storage_modify() {
+        let account = Account {
+            client: 1,
+            available: 20.0,
+            held: 0.0,
+            total: 0.0,
+            locked: false,
+        };
+
+        let db = AccountStorage::new();
+        let exists = db.exists(1);
+        assert!(!exists, "account should be empty");
+
+        db.insert(account.clone());
+
+        let acct: Account = db.read(1, |acct| acct.unwrap().clone());
+
+        assert!(
+            acct.available == account.available,
+            "created client id and fetched client id are not equal"
+        );
+
+        let updated = db.modify(acct.client, |acc| {
+            let a = acc.unwrap();
+            a.available = 25.0;
+
+            return *a;
+        });
+
+        db.insert(updated);
+
+        assert!(
+            updated.available == 25.0,
+            "account not equal after modification; expect {}, got {}",
+            25.0,
+            updated.available,
         );
     }
 }
